Splited UpdatePopulation in two: model and result class. Added functions to modify settings of UpP model
